requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Pillow==10.0.0
markdown==3.5
bleach==6.0.0
Authlib==1.3.0